requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
SQLAlchemy>=2.0,<3.0
pydantic>=2.9.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0
pytest>=8.0.0
httpx>=0.27.0